[#164238813] separate of concerns
