+++ routes/articles.js
@@ -2,12 +2,14 @@ import { Router } from "express";
 import Article from "../controllers/articlesController";
 import articleValidator from "../middlewares/articleValidator";
 import ArticleLikeController from "../controllers/articleLikesController";
+import BookmarkController from "../controllers/bookmarkController";
+import ReportArticleController from "../controllers/reportArticleController";
+import ShareController from '../controllers/shareController';
 import checkAuth from "../middlewares/checkAuth";
 
 const article = Router();
 
-
-article.post("/articles/:slug/bookmark", checkAuth, Article.bookmark);
+article.post("/articles/:slug/bookmark", checkAuth, BookmarkController.bookmark);
 article.post("/articles/:slug/likes", ArticleLikeController.like);
 article.post("/articles/:slug/dislikes", ArticleLikeController.dislike);
 article.get("/articles/:slug/likes", ArticleLikeController.getArticleLikes);
@@ -18,22 +20,25 @@ article
     "/articles/:slug/share/twitter",
     checkAuth,
     articleValidator.validateArticle,
-    Article.shareOnTwitter
+    ShareController.shareOnTwitter
   )
   .get(
     "/articles/:slug/share/fb",
     checkAuth,
     articleValidator.validateArticle,
-    Article.shareOnFacebook
+    ShareController.shareOnFacebook
   )
   .get(
     "/articles/:slug/share/linkedIn",
     checkAuth,
     articleValidator.validateArticle,
-    Article.shareOnLinkedin
+    ShareController.shareOnLinkedin
   )
-  .get("/articles/:slug/share/email", checkAuth, Article.shareOnEmail)
+  .get("/articles/:slug/share/email", checkAuth, ShareController.shareOnEmail)
   .get("/articles/:slug", Article.findOneArticle);
-article.post("/articles/:slug/bookmark", Article.bookmark);
-article.put("/articles/:slug/report", articleValidator.validateArticle, Article.reportArticle);
+article.put(
+  "/articles/:slug/report",
+  articleValidator.validateArticle,
+  ReportArticleController.reportArticle
+);
 export default article;
